[Fix] LikeRepository - id 자료형 수정

+++ src/main/java/com/woowahan/recipe/repository/LikeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LikeRepository extends JpaRepository<LikeEntity, Integer> {  // FIXME: 2023-01-17  ID 자료형 수정
+public interface LikeRepository extends JpaRepository<LikeEntity, Long> {
 }
